+++ CHANGES.md
@@ -4,6 +4,8 @@ Changelog for openvario-shell
 0.3 (unreleased)
 ----------------
 
+- Initial API to support 3rd party apps
+
 - Do not write to xcsoar profile on startup if no changes are made